fix: update twine to address importlib-metadata issue
Update twine to version 5.1.1 or higher to address a bug stemming from
twine's reliance on deprecated behavior, which was removed from
importlib-metadata in version 8.0.0.
For more details, see:
- https://github.com/python/importlib_metadata/pull/491/commits/a970a491b56a3bf529821ce21867af4573cf2e0d
- https://github.com/pypa/twine/issues/977

The plugin currently uses twine version 4.0.2, which depends on the
latest version of importlib-metadata (>= 3.6).
Consequently, it is impacted by the breaking changes introduced in
importlib-metadata version 8.0.0.

This issue has been resolved in twine version 5.1.1, as detailed here:
https://github.com/pypa/twine/pull/1124.